Refactors ensures column sizes
